+++ packages/engine-multi/src/classes/ExecutionContext.ts
@@ -7,6 +7,7 @@ import type {
   ExecutionContextOptions,
 } from '../types';
 import type { Logger } from '@openfn/logger';
+import loadVersions from '../util/load-versions';
 
 /**
  * The ExeuctionContext class wraps an event emitter with some useful context
@@ -21,6 +22,7 @@ export default class ExecutionContext extends EventEmitter {
   logger: Logger;
   callWorker: CallWorker;
   options: ExecutionContextOptions;
+  versions = {};
 
   constructor({
     state,
@@ -33,12 +35,13 @@ export default class ExecutionContext extends EventEmitter {
     this.callWorker = callWorker;
     this.state = state;
     this.options = options;
+    this.versions = loadVersions();
   }
 
   // override emit to add the workflowId to all events
   // @ts-ignore
   emit(event: string, payload: any) {
     payload.workflowId = this.state.id;
-    super.emit(event, payload);
+    return super.emit(event, payload);
   }
 }