+++ packages/engine-multi/src/types.ts
@@ -54,6 +54,7 @@ export interface ExecutionContext extends EventEmitter {
   state: WorkflowState;
   logger: Logger;
   callWorker: CallWorker;
+  versions: Versions;
 
   emit<T extends ExternalEvents>(
     event: T,
@@ -67,15 +68,12 @@ export interface EngineAPI extends EventEmitter {
   purge?: () => void;
 }
 
-export interface RuntimeEngine extends EventEmitter {
-  //id: string // human readable instance id
-  // actually I think the id is on the worker, not the engine
+export interface RuntimeEngine {
+  version: string;
 
   // TODO should return an unsubscribe hook
   listen(attemptId: string, listeners: any): void;
 
-  // TODO return a promise?
-  // Kinda convenient but not actually needed
   execute(
     plan: ExecutionPlan,
     options?: Partial<EngineOptions>
@@ -83,5 +81,14 @@ export interface RuntimeEngine extends EventEmitter {
 
   destroy(): void;
 
+  on: (evt: string, fn: (...args: any[]) => void) => void;
+
   // TODO my want some maintenance APIs, like getStatus. idk
 }
+
+export type Versions = {
+  node: string;
+  engine: string;
+  compiler: string;
+  [adaptor: string]: string;
+};