+++ packages/ws-worker/src/types.d.ts
@@ -1,5 +1,6 @@
 import { SanitizePolicies } from '@openfn/logger';
 import type { Channel as PhxChannel } from 'phoenix';
+import type { ExecutionPlan } from '@openfn/runtime';
 
 export { Socket };
 
@@ -41,7 +42,8 @@ export type Node = {
   id: string;
   body?: string;
   adaptor?: string;
-  credential_id?: any; // TODO tighten this up, string or object
+  credential?: object;
+  credential_id?: string;
   type?: 'webhook' | 'cron'; // trigger only
   state?: any; // Initial state / defaults
 };